Finished Metric's classes

+++ src/main/java/trainning/tdd/metric/Foot.java
@@ -4,18 +4,18 @@ package trainning.tdd.metric;
  * Created by chenzi on 2019/9/30.
  */
 public class Foot {
-    private final int footNumber;
+    private final double footNumber;
 
-    public Foot(int footNumber) {
+    public Foot(double footNumber) {
         this.footNumber = footNumber;
     }
 
-    public int toInches() {
-        return footNumber * 12;
+    public double toInches() {
+        return footNumber * 12.0;
     }
 
-    public int toYards() {
-        return footNumber / 3;
+    public double toYards() {
+        return footNumber / 3.0;
     }
 
 }
